+++ NEWS.md
@@ -1,6 +1,9 @@
 # Development
 
-* `prepare_geo`: convert non-character ID column in gpl data to character to avoid the ID column is like numeric 1, 2, 3, 4...
+* `prepare_geo`: 
+  
+1. convert non-character ID column in gpl data to character to avoid the ID column is like numeric 1, 2, 3, 4...
+2. read potential expression data from supplementary files...
 
 # easybio 1.0.1